feat: :monocle_face: dynamically encode verbs
